+++ out/glsl/export.js
@@ -29,6 +29,7 @@ class GlslExport {
                 textures[key] = texture;
             }
         }
+        const extensions = options.extensions;
         // const uniforms = JSON.stringify(options.uniforms, null, '\t');
         GlslExport.selectFolder().then((outputPath) => {
             const resourcePath = vscode.Uri.file(path.join(extensionPath, 'resources', 'export')).fsPath;
@@ -39,7 +40,7 @@ class GlslExport {
                 GlslExport.copyFolder(path.join(resourcePath, 'css'), path.join(outputPath, 'css')),
                 GlslExport.copyFolder(path.join(resourcePath, 'img'), path.join(outputPath, 'img')),
                 GlslExport.copyFile(path.join(extensionPath, 'node_modules', 'stats.js', 'build', 'stats.min.js'), path.join(outputPath, 'js', 'stats.min.js')),
-                GlslExport.copyFile(path.join(extensionPath, 'node_modules', 'glsl-canvas-js', 'dist', 'glsl-canvas.min.js'), path.join(outputPath, 'js', 'glsl-canvas.min.js')),
+                GlslExport.copyFile(path.join(extensionPath, 'node_modules', 'glsl-canvas-js', 'dist', 'umd', 'glsl-canvas.min.js'), path.join(outputPath, 'js', 'glsl-canvas.min.js')),
                 GlslExport.writeFile(options.fragment, path.join(outputPath, 'shaders', 'fragment.glsl')),
                 GlslExport.writeFile(`<!DOCTYPE html>
 	<html>
@@ -73,7 +74,11 @@ class GlslExport {
 		Events: load, error, textureError, render, over, out, move, click
 		Methods: load, on, pause, play, toggle, setTexture, setUniform, setUniforms, destroy
 		*/
-		var glsl = GlslCanvas.of(document.querySelector('.glsl-canvas'));
+		var options = {
+			extensions: ${JSON.stringify(extensions)},
+		};
+		var canvas = document.querySelector('.glsl-canvas');
+		var glsl = new glsl.Canvas(canvas, options);
 		glsl.setUniforms(${JSON.stringify(uniforms)});${Object.keys(textures).map(x => `
 		glsl.setTexture('u_texture_${x}', '${textures[x]}', {
 			filtering: 'mipmap',
@@ -139,7 +144,12 @@ class GlslExport {
     static copyFolder(src, dest) {
         return new Promise((resolve, reject) => {
             GlslExport.readFiles(src).then((files) => {
-                Promise.all(files.map(x => GlslExport.copyFile(path.join(src, x), path.join(dest, x)))).then(files => {
+                Promise.all(files.filter(x => {
+                    const filePath = path.join(src, x);
+                    return fs.existsSync(filePath) && fs.lstatSync(filePath).isFile();
+                }).map(x => {
+                    return GlslExport.copyFile(path.join(src, x), path.join(dest, x));
+                })).then(files => {
                     resolve(files);
                 }, error => {
                     reject(error);
@@ -304,6 +314,6 @@ class GlslExport {
         }
     }
 }
-GlslExport.npm = false;
 exports.default = GlslExport;
+GlslExport.npm = false;
 //# sourceMappingURL=export.js.map