+++ out/glsl/options.js
@@ -13,6 +13,7 @@ class GlslOptions {
         this.uniforms = config['uniforms'] || {};
         this.timeout = config['timeout'] || 0;
         this.textures = Object.assign({}, config['textures'] || {});
+        this.extensions = config['extensions'] || [];
         const folder = vscode.workspace ? vscode.workspace.rootPath : null;
         if (folder) {
             if (this.fragment !== '') {