+++ src/glsl/panel.ts
@@ -5,26 +5,32 @@ import * as vscode from 'vscode';
 import { Uri } from 'vscode';
 import GlslOptions from './options';
 
+const DESERIALIZE_PANEL: boolean = true;
+
 export default class GlslPanel {
 
 	public static current: GlslPanel | undefined;
 	public static readonly viewType = 'glslCanvas';
-	private readonly panel: vscode.WebviewPanel;
-	private readonly extensionPath: string;
-	private disposables: vscode.Disposable[] = [];
-	private _onMessage: Function;
+	private readonly panel_: vscode.WebviewPanel;
+	private readonly extensionPath_: string;
+	private disposables_: vscode.Disposable[] = [];
+	private onMessage_: Function;
+	private onMessageDisposable_: vscode.Disposable;
 
 	set onMessage(onMessage: Function) {
-		this._onMessage = onMessage;
-		this.panel.webview.onDidReceiveMessage(message => {
+		this.onMessage_ = onMessage;
+		if (this.onMessageDisposable_) {
+			this.onMessageDisposable_.dispose();
+		}
+		this.onMessageDisposable_ = this.panel_.webview.onDidReceiveMessage(message => {
 			if (typeof onMessage === 'function') {
 				onMessage(message);
 			}
-		}, null, this.disposables);
+		}, null, this.disposables_);
 	}
 
 	get onMessage(): Function {
-		return this._onMessage;
+		return this.onMessage_;
 	}
 
 	private constructor(
@@ -32,53 +38,65 @@ export default class GlslPanel {
 		extensionPath: string,
 		onMessage?: Function,
 	) {
-		this.panel = panel;
-		this.extensionPath = extensionPath;
-		this.render();
-		this.panel.onDidDispose(() => this.dispose(), null, this.disposables);
-		this.panel.onDidChangeViewState(e => {
-            /*
-            if (this.panel.visible) {
-                console.log('onDidChangeViewState', e);
-                // this.render();
-            }
-            */
-		}, null, this.disposables);
+		this.panel_ = panel;
+		this.extensionPath_ = extensionPath;
+		this.panel_.onDidDispose(() => this.dispose(), null, this.disposables_);
+		/*
+		// Update the content based on view changes
+		// cause an unneeded reload of the canvas
+		this.panel_.onDidChangeViewState(e => {
+			if (this.panel_.visible) {
+				// console.log('onDidChangeViewState', e);
+				this.render();
+			}
+		}, null, this.disposables_);
+		*/
 		this.onMessage = onMessage;
+		this.render();
+		// console.log('GlslPanel');
 	}
 
 	public static createOrShow(extensionPath: string, onMessage?: Function): GlslPanel {
 		const viewColumn = vscode.window.activeTextEditor ? vscode.ViewColumn.Beside : vscode.ViewColumn.One;
 		if (GlslPanel.current) {
-			GlslPanel.current.onMessage = onMessage;
-			GlslPanel.current.panel.reveal(viewColumn, true);
-			return GlslPanel.current;
+			if (DESERIALIZE_PANEL) {
+				GlslPanel.current.onMessage = onMessage;
+				GlslPanel.current.panel_.reveal(viewColumn, true);
+				return GlslPanel.current;
+			} else {
+				GlslPanel.current.dispose();
+			}
 		}
-		const localResourceRoots = [];
-		localResourceRoots.push(
-			vscode.Uri.file(path.join(extensionPath, 'resources'))
-		);
-		if (vscode.workspace && vscode.workspace.rootPath) {
-			localResourceRoots.push(
-				vscode.Uri.file(vscode.workspace.rootPath)
-			);
-		}
+		const localResourceRoots = this.getLocalResourceRoots(extensionPath);
 		const panel = vscode.window.createWebviewPanel(
 			GlslPanel.viewType,
-			'GlslCanvas', {
-			viewColumn: viewColumn,
-			preserveFocus: true,
-		}, {
-			enableScripts: true,
-			retainContextWhenHidden: true, // !!!
-			localResourceRoots: localResourceRoots,
-		}
+			'GlslCanvas',
+			{
+				viewColumn: viewColumn,
+				preserveFocus: true,
+			},
+			{
+				enableScripts: true,
+				retainContextWhenHidden: true, // !!!
+				localResourceRoots: localResourceRoots,
+			}
 		);
 		GlslPanel.current = new GlslPanel(panel, extensionPath, onMessage);
 		return GlslPanel.current;
 	}
 
+	public static getLocalResourceRoots(extensionPath: string): vscode.Uri[] {
+		const localResourceRoots = [];
+		localResourceRoots.push(vscode.Uri.file(path.join(extensionPath, 'resources')));
+		if (vscode.workspace && vscode.workspace.rootPath) {
+			localResourceRoots.push(vscode.Uri.file(vscode.workspace.rootPath));
+		}
+		// console.log(localResourceRoots);
+		return localResourceRoots;
+	}
+
 	public static revive(panel: vscode.WebviewPanel, extensionPath: string, onMessage?: Function) {
+		// panel.webview.options.localResourceRoots.concat(this.getLocalResourceRoots(extensionPath));
 		GlslPanel.current = new GlslPanel(panel, extensionPath, onMessage);
 	}
 
@@ -88,9 +106,15 @@ export default class GlslPanel {
 		}
 	}
 
+	public static render(uri: vscode.Uri) {
+		if (GlslPanel.current) {
+			GlslPanel.current.render(uri);
+		}
+	}
+
 	public static rebuild(onMessage?: Function) {
 		if (GlslPanel.current) {
-			const extensionPath = GlslPanel.current.extensionPath;
+			const extensionPath = GlslPanel.current.extensionPath_;
 			GlslPanel.dispose();
 			GlslPanel.createOrShow(extensionPath, onMessage);
 		}
@@ -101,9 +125,9 @@ export default class GlslPanel {
             /*
             const viewColumn = vscode.window.activeTextEditor ? vscode.ViewColumn.Beside : vscode.ViewColumn.One;
             console.log('reveal', vscode.window.activeTextEditor, viewColumn);
-            GlslPanel.current.panel.reveal(viewColumn, true);
+            GlslPanel.current.panel_.reveal(viewColumn, true);
             */
-			GlslPanel.current.panel.reveal(vscode.ViewColumn.Beside, true);
+			GlslPanel.current.panel_.reveal(vscode.ViewColumn.Beside, true);
 			GlslPanel.current.update(uri);
 		}
 	}
@@ -115,19 +139,19 @@ export default class GlslPanel {
 	}
 
 	public dispose() {
-		GlslPanel.current = undefined;
-		this.panel.dispose();
-		while (this.disposables.length) {
-			const x = this.disposables.pop();
+		while (this.disposables_.length) {
+			const x = this.disposables_.pop();
 			if (x) {
 				x.dispose();
 			}
 		}
+		this.panel_.dispose();
+		GlslPanel.current = undefined;
 	}
 
 	public update(uri: Uri) {
 		const options = new GlslOptions();
-		this.panel.webview.postMessage(options.serialize()).then(
+		this.panel_.webview.postMessage(options.serialize()).then(
 			(success) => {
 				// console.log('GlslPanel.update.success');
 			},
@@ -138,26 +162,39 @@ export default class GlslPanel {
 		);
 	}
 
-	private render() {
-		this.panel.title = 'glslCanvas';
-		this.panel.webview.html = this.getPanelWebviewHtml();
+	private render(uri?: vscode.Uri) {
+		this.panel_.title = 'glslCanvas';
+		this.panel_.webview.html = this.getPanelWebviewHtml(uri);
 	}
 
-	private getPanelWebviewHtml(): string {
+	private getPanelWebviewHtml(uri?: vscode.Uri): string {
 		const config = vscode.workspace.getConfiguration('editor');
 		const options = new GlslOptions();
+		const nonce = this.getNonce();
+		const cspSource = this.getCspSource();
 		// console.log('getPanelWebviewHtml', config, options);
-		const content = `
+		const content = `<!DOCTYPE html>
+<html lang="en">
 <head>
-	<link href="${this.getResource('fonts/styles.css')}" rel="stylesheet">
-	<!-- <link href="${this.getResource('css/vendors.min.css')}" rel="stylesheet"> -->
+	<meta charset="UTF-8">
+	<!--
+	<meta http-equiv="Content-Security-Policy" content="
+	default-src 'none';
+	script-src ${cspSource} 'self' 'nonce-${nonce}';
+	style-src ${cspSource} 'self' 'nonce-${nonce}';
+	img-src ${cspSource} 'self' https:;
+	font-src ${cspSource} 'self';
+	connect-src 'self'; " />
+	-->
+	<meta name="viewport" content="width=device-width, initial-scale=1.0">
+	<link nonce="${nonce}" href="${this.getResource('fonts/styles.css')}" rel="stylesheet">
+	<!-- <link nonce="${nonce}" href="${this.getResource('css/vendors.min.css')}" rel="stylesheet"> -->
 	<style>
 		html, body { font-family: ${config.fontFamily}; font-weight: ${config.fontWeight}; font-size: ${config.fontSize}; };
 	</style>
-	<script nonce="${this.getNonce()}" src="${this.getResource('js/vendors.min.js')}"></script>
-	<link href="${this.getResource('css/app.min.css')}" rel="stylesheet"/>
+	<link nonce="${nonce}" href="${this.getResource('css/app.min.css')}" rel="stylesheet"/>
 </head>
-<script>
+<script nonce="${nonce}">
 	var options = ${options.serialize()};
 </script>
 <body class="idle">
@@ -173,17 +210,24 @@ export default class GlslPanel {
 	<div class="errors"></div>
 	<div class="welcome"><div class="welcome-content" unselectable><p>There's no active .glsl editor</p><button class="btn-create"><span>create one</span></button></div></div>
 	<div class="missing"><div class="missing-content" unselectable><p>Oops. There was a problem with WebGL.</p></div></div>
-	<script nonce="${this.getNonce()}" src="${this.getResource('js/app.min.js')}"></script>
+	<script nonce="${nonce}" src="${this.getResource('js/vendors.min.js')}"></script>
+	<script nonce="${nonce}" src="${this.getResource('js/app.min.js')}"></script>
 </body>
-        `;
+</html>`;
 		return content;
 	}
 
+	private getCspSource(): string {
+		return this.panel_.webview.cspSource;
+	}
+
 	private getResource(resource: string): vscode.Uri {
-		const file = vscode.Uri.file(path.join(this.extensionPath, 'resources', resource));
-		const vscodeResource = file.with({ scheme: 'vscode-resource' });
-		// const panel = this.panel;
-		// const vscodeResource = panel.webview.toWebviewResource(imagePath);
+		const fileOnDisk = vscode.Uri.file(
+			path.join(this.extensionPath_, 'resources', resource)
+		);
+		// const vscodeResource = fileOnDisk.with({ scheme: 'vscode-resource' });
+		const panel = this.panel_;
+		const vscodeResource = panel.webview.asWebviewUri(fileOnDisk);
 		return vscodeResource;
 	}