+++ CHANGELOG.md
@@ -3,9 +3,16 @@ All notable changes to the "vscode-glsl-canvas" extension will be documented in 
 
 ---
 
+## [0.2.9] - 2020-03-29
+### Added
+- WebGL extensions option support.
+- Inizialization protocol error.
+
+---
+
 ## [0.2.8] - 2020-01-01
 ### Added
-- texture querystring options.
+- Texture querystring options.
 
 ---